create routes folder, and Refactor router import in App.js
