+++ src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Router from './router';
+import Router from './routes/Router';
 import './styles/App.css';
 
 const App = () => {
